models/pycaret/requirements.txt
# Core dependencies
pycaret==3.3.2
pandas>=1.3.5
numpy>=1.21.0
scikit-learn>=1.0.2
matplotlib>=3.5.1
seaborn>=0.11.2

# ML libraries used by PyCaret
xgboost>=1.5.0
lightgbm>=3.3.2
catboost>=1.0.6

# Experiment tracking
mlflow>=2.0.0

# Utilities
PyYAML>=6.0
joblib>=1.1.0
tqdm>=4.64.0

# Optional: for parallel processing
ray>=2.0.0

# Optional: for hyperparameter optimization
optuna>=3.0.0
hyperopt>=0.2.7

# Optional: for calibration
netcal>=1.3.0

# Optional: for feature selection
shap>=0.41.0 